Create code_R
Taking as input data the pixel matrix in "txt format", this code calculates the lacunarity curve and the lacunarity exponent through the "Gliding Box" algorithm. The pixel matrix can come from any type of digital image or microscopy. The "L" parameter is the dimension of the matrix and must be indicated by an integer. The "threshold" parameter depends on the characteristics of the data and can be obtained externally using some automated method such as the OTSU algorithm or Artificial Intelligence, however, we suggest as default the average of the pixel matrix data. The "m" parameter represents the values of the box size, which can be a sequence of integers "m = 1,2,4,8,16,32,64,128". This sequence can be changed and extended to receive other values as long as the "lacunarity" function is adjusted for this. A test matrix (surf.txt) for evaluating the code is also provided.